Finally got websockets working

+++ database/migrations/2019_12_28_205807_create_websockets_statistics_entries_table.php
@@ -12,11 +12,11 @@ class CreateWebSocketsStatisticsEntriesTable extends Migration
     public function up()
     {
         Schema::create('websockets_statistics_entries', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('app_id');
-            $table->integer('peak_connection_count');
-            $table->integer('websocket_message_count');
-            $table->integer('api_message_count');
+            $table->bigInteger('peak_connection_count');
+            $table->bigInteger('websocket_message_count');
+            $table->bigInteger('api_message_count');
             $table->nullableTimestamps();
         });
     }
